// components/about/about.types.ts
export interface CompanyStat {
  label: string;
  value: string;
}

export interface CoreValue {
  title: string;
  description: string;
  icon: string;
}

export interface Certification {
  title: string;
  image: string;
}

export interface LeadershipMember {
  name: string;
  designation: string;
  image: string;
  description: string;
}